requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
drf-spectacular>=0.27,<1.0
django-filter>=24.2,<26.0
psycopg2-binary>=2.9,<3.0
